chore: better icons
