+++ src/svg-icons/common/Attachment.tsx
@@ -44,3 +44,17 @@ export const DocumentAttachmentIcon = (): JSX.Element => {
     </Svg>
   );
 };
+
+export const AudioIcon = (): JSX.Element => {
+  return (
+    <Svg width="16" height="16" viewBox="0 0 16 16" fill="none">
+      <Path
+        d="M12.6654 6.66634V7.99967C12.6654 9.23735 12.1737 10.4243 11.2985 11.2995C10.4234 12.1747 9.23637 12.6663 7.9987 12.6663M7.9987 12.6663C6.76102 12.6663 5.57404 12.1747 4.69887 11.2995C3.8237 10.4243 3.33203 9.23735 3.33203 7.99967V6.66634M7.9987 12.6663V14.6663M7.9987 1.33301C7.46826 1.33301 6.95956 1.54372 6.58448 1.91879C6.20941 2.29387 5.9987 2.80257 5.9987 3.33301V7.99967C5.9987 8.53011 6.20941 9.03882 6.58448 9.41389C6.95956 9.78896 7.46826 9.99967 7.9987 9.99967C8.52913 9.99967 9.03784 9.78896 9.41291 9.41389C9.78798 9.03882 9.9987 8.53011 9.9987 7.99967V3.33301C9.9987 2.80257 9.78798 2.29387 9.41291 1.91879C9.03784 1.54372 8.52913 1.33301 7.9987 1.33301Z"
+        stroke="#8D8D8D"
+        strokeWidth="1.33333"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+    </Svg>
+  );
+};
