--- src/svg-icons/common/Outgoing.tsx
@@ -1,16 +1,0 @@
-import React from 'react';
-import Svg, { Path } from 'react-native-svg';
-
-export const OutgoingIcon = () => {
-  return (
-    <Svg width="16" height="16" viewBox="0 0 20 20" fill="none">
-      <Path
-        d="M7.50065 11.6663L3.33398 7.49967M3.33398 7.49967L7.50065 3.33301M3.33398 7.49967H13.334C14.218 7.49967 15.0659 7.85086 15.691 8.47598C16.3161 9.10111 16.6673 9.94895 16.6673 10.833V16.6663"
-        stroke="#646464"
-        strokeWidth="1.66667"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-    </Svg>
-  );
-};
